Query processing:  minor bug fix in a test.

+++ src/Samples/QueryProcessingTest/SqlBugsTest.cs
@@ -64,7 +64,7 @@ namespace QueryProcessingTest {
             Trace.Assert(row is IObjectView);
             Trace.Assert(row2 is IObjectView);
             Trace.Assert((row as IObjectView).GetObject(0).GetObjectNo() == (row2 as IObjectView).GetObject(0).GetObjectNo());
-            Trace.Assert((row as IObjectView).GetInt16(1) == (row2 as IObjectView).GetInt64(1));
+            Trace.Assert((row as IObjectView).GetInt64(1) == (row2 as IObjectView).GetInt64(1));
             HelpMethods.LogEvent("Finished some tests on variables and case insensitivity");
             HelpMethods.LogEvent("Start testing queries on comparison bug");
             TestComparison();
